initial round of amd64 fixes

+++ code/Ratl/ratl_common.h
@@ -789,11 +789,11 @@ namespace storage
 		// this is so node support does not need to be added to the primitive containers
 		static NODE & node(TValue &v)
 		{
-			return *(NODE *)((unsigned char *)(&v)+int(&((TStorage *)0)->nodeData)-int(&((TStorage *)0)->value));
+			return *(NODE *)((unsigned char *)(&v)+size_t(&((TStorage *)0)->nodeData)-size_t(&((TStorage *)0)->value));
 		}
 		static const NODE & node(const TValue &v)
 		{
-			return *(const NODE *)((unsigned char *)(&v)+int(&((TStorage *)0)->nodeData)-int(&((TStorage *)0)->value));
+			return *(const NODE *)((unsigned char *)(&v)+size_t(&((TStorage *)0)->nodeData)-size_t(&((TStorage *)0)->value));
 		}
 		static void swap(TStorage *s1,TStorage *s2)
 		{
@@ -803,8 +803,8 @@ namespace storage
 		static int pointer_to_index(const void *s1,const void *s2)
 		{
 			return 
-				((TStorage *)(((unsigned char *)s1)-int(&((TStorage *)0)->value))) - 
-				((TStorage *)(((unsigned char *)s2)-int(&((TStorage *)0)->value)));
+				((TStorage *)(((unsigned char *)s1)-size_t(&((TStorage *)0)->value))) - 
+				((TStorage *)(((unsigned char *)s2)-size_t(&((TStorage *)0)->value)));
 		}	
 	};
 
@@ -884,11 +884,11 @@ namespace storage
 		// this is so node support does not need to be added to the primitive containers
 		static NODE & node(TValue &v)
 		{
-			return *(NODE *)((unsigned char *)(&v)+int(&((TStorage *)0)->nodeData)-int(&((TStorage *)0)->value));
+			return *(NODE *)((unsigned char *)(&v)+size_t(&((TStorage *)0)->nodeData)-size_t(&((TStorage *)0)->value));
 		}
 		static const NODE & node(const TValue &v)
 		{
-			return *(const NODE *)((unsigned char *)(&v)+int(&((TStorage *)0)->nodeData)-int(&((TStorage *)0)->value));
+			return *(const NODE *)((unsigned char *)(&v)+size_t(&((TStorage *)0)->nodeData)-size_t(&((TStorage *)0)->value));
 		}
 		static void swap(TStorage *s1,TStorage *s2)
 		{
@@ -900,8 +900,8 @@ namespace storage
 		static int pointer_to_index(const void *s1,const void *s2)
 		{
 			return 
-				((TStorage *)(((unsigned char *)s1)-int(&((TStorage *)0)->value))) - 
-				((TStorage *)(((unsigned char *)s2)-int(&((TStorage *)0)->value)));
+				((TStorage *)(((unsigned char *)s1)-size_t(&((TStorage *)0)->value))) - 
+				((TStorage *)(((unsigned char *)s2)-size_t(&((TStorage *)0)->value)));
 		}	
 	};
 	template<class T,int SIZE,int MAX_CLASS_SIZE,class NODE>
@@ -975,11 +975,11 @@ namespace storage
 		// this is so node support does not need to be added to the primitive containers
 		static NODE & node(TValue &v)
 		{
-			return *(NODE *)((unsigned char *)(&v)+int(&((TStorage *)0)->nodeData)-int(&((TStorage *)0)->value));
+			return *(NODE *)((unsigned char *)(&v)+size_t(&((TStorage *)0)->nodeData)-size_t(&((TStorage *)0)->value));
 		}
 		static const NODE & node(const TValue &v)
 		{
-			return *(const NODE *)((unsigned char *)(&v)+int(&((TStorage *)0)->nodeData)-int(&((TStorage *)0)->value));
+			return *(const NODE *)((unsigned char *)(&v)+size_t(&((TStorage *)0)->nodeData)-size_t(&((TStorage *)0)->value));
 		}
 		// this is a bit suspicious, we are forced to do a memory swap, and for a class, that, say
 		// stores a pointer to itself, it won't work right
@@ -991,8 +991,8 @@ namespace storage
 		static int pointer_to_index(const void *s1,const void *s2)
 		{
 			return 
-				((TStorage *)(((unsigned char *)s1)-int(&((TStorage *)0)->value))) - 
-				((TStorage *)(((unsigned char *)s2)-int(&((TStorage *)0)->value)));
+				((TStorage *)(((unsigned char *)s1)-size_t(&((TStorage *)0)->value))) - 
+				((TStorage *)(((unsigned char *)s2)-size_t(&((TStorage *)0)->value)));
 		}	
 		template<class CAST_TO>
 		static CAST_TO *verify_alloc(CAST_TO *p)
