+++ code/cgame/cg_main.cpp
@@ -93,7 +93,8 @@ This must be the very first function compiled into the .q3vm file
 ================
 */
 extern "C" {
-int vmMain( int command, int arg0, int arg1, int arg2, int arg3, int arg4, int arg5, int arg6, int arg7 ) {
+intptr_t vmMain( intptr_t command, intptr_t arg0, intptr_t arg1, intptr_t arg2,
+    intptr_t arg3, intptr_t arg4, intptr_t arg5, intptr_t arg6, intptr_t arg7 ) {
 	centity_t		*cent;
 
 	switch ( command ) {
