+++ code/qcommon/MiniHeap.h
@@ -16,9 +16,9 @@ public:
 void ResetHeap()
 {
 #if _DEBUG
-	if ((int)mCurrentHeap - (int)mHeap>mMaxAlloc)
+	if ((size_t)mCurrentHeap - (size_t)mHeap>mMaxAlloc)
 	{
-		mMaxAlloc=(int)mCurrentHeap - (int)mHeap;
+		mMaxAlloc=(size_t)mCurrentHeap - (size_t)mHeap;
 	}
 #endif
 	mCurrentHeap = mHeap;
@@ -50,7 +50,7 @@ CMiniHeap(int size)
 // give me some space from the heap please
 char *MiniHeapAlloc(int size)
 {
-	if (size < (mSize - ((int)mCurrentHeap - (int)mHeap)))
+	if (size < (mSize - ((size_t)mCurrentHeap - (size_t)mHeap)))
 	{
 		char *tempAddress =  mCurrentHeap;
 		mCurrentHeap += size;
